Product admin - Securing the views

+++ watches/views.py
@@ -1,5 +1,6 @@
 from django.shortcuts import render, redirect, reverse, get_object_or_404
 from django.contrib import messages
+from django.contrib.auth.decorators import login_required
 from django.db.models import Q
 from django.db.models.functions import Lower
 
@@ -73,8 +74,13 @@ def watch_detail(request, watch_id):
     return render(request, 'watches/watch_detail.html', context)
 
 
+@login_required
 def add_watch(request):
     """ Add a watch to the store """
+    if not request.user.is_superuser:
+        messages.error(request, 'Sorry, only store owner can do that.')
+        return redirect(reverse('home'))
+
     if request.method == 'POST':
         form = ProductForm(request.POST, request.FILES)
         if form.is_valid():
@@ -95,8 +101,13 @@ def add_watch(request):
     return render(request, template, context)
 
 
+@login_required
 def edit_watch(request, watch_id):
     """ Edit a watch currently in store """
+    if not request.user.is_superuser:
+        messages.error(request, 'Sorry, only store owner can do that.')
+        return redirect(reverse('home'))
+
     watch = get_object_or_404(Watch, pk=watch_id)
     if request.method == 'POST':
         form = ProductForm(request.POST, request.FILES, instance=watch)
@@ -120,8 +131,13 @@ def edit_watch(request, watch_id):
     return render(request, template, context)
 
 
+@login_required
 def delete_watch(request, watch_id):
     """ Delete a watch currently in store """
+    if not request.user.is_superuser:
+        messages.error(request, 'Sorry, only store owner can do that.')
+        return redirect(reverse('home'))
+
     watch = get_object_or_404(Watch, pk=watch_id)
     watch.delete()
     messages.success(request, 'Watch successfully deleted.')
